Refactor projects page
Each project lives in its own .md file in src/content/projects. The
projects are then displayed as gridded cards on the projects page.

We're not importing from notion at this point because the information is
not organized well enough to display, and also so we can format the
description with markdown.